+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add a full set of bench positions for 'bench' command
 * When in a tablebase position, report the tablebase PV line
 * Don't panic when SyzygyPath is set but is empty
+* Rewrite the FEN parser without using `nom`
 * Use Rust 1.85 and Edition 2024
 
 ## [5.1]